Add jinja2 template
Add jinja2 and use it to render a template for the root method. Used a short helper function to do the rendering.

Added tests that do just enough to make me feel confident that the render is working, and that the index method is using the correct template and arguments. I waffled about testing that the index.html template rendered to the HTML that I expected and may still revist this later, but seemed excessive for now.

Also ran black auto-formatting on all the python files.